requirements.txt
# Django Core
Django==4.2.7
djangorestframework==3.14.0

# Authentication
djangorestframework-simplejwt==5.3.0

# Database
mysqlclient==2.2.0

# CORS
django-cors-headers==4.3.0

# Environment Variables
python-decouple==3.8

# Utilities
Pillow>=10.2.0
python-dateutil==2.8.2

# Payment Gateway
razorpay==1.4.1

# PDF Generation
weasyprint==60.1
